// SDKs/shotstack.io/v1/typescript/src/sdk/models/shared/transition.ts
export enum TransitionInEnum {
    Fade = "fade",
    FadeSlow = "fadeSlow",
    FadeFast = "fadeFast",
    Reveal = "reveal",
    RevealSlow = "revealSlow",
    RevealFast = "revealFast",
    WipeLeft = "wipeLeft",
    WipeLeftSlow = "wipeLeftSlow",
    WipeLeftFast = "wipeLeftFast",
    WipeRight = "wipeRight",
    WipeRightSlow = "wipeRightSlow",
    WipeRightFast = "wipeRightFast",
    SlideLeft = "slideLeft",
    SlideLeftSlow = "slideLeftSlow",
    SlideLeftFast = "slideLeftFast",
    SlideRight = "slideRight",
    SlideRightSlow = "slideRightSlow",
    SlideRightFast = "slideRightFast",
    SlideUp = "slideUp",
    SlideUpSlow = "slideUpSlow",
    SlideUpFast = "slideUpFast",
    SlideDown = "slideDown",
    SlideDownSlow = "slideDownSlow",
    SlideDownFast = "slideDownFast",
    CarouselLeft = "carouselLeft",
    CarouselLeftSlow = "carouselLeftSlow",
    CarouselLeftFast = "carouselLeftFast",
    CarouselRight = "carouselRight",
    CarouselRightSlow = "carouselRightSlow",
    CarouselRightFast = "carouselRightFast",
    CarouselUp = "carouselUp",
    CarouselUpSlow = "carouselUpSlow",
    CarouselUpFast = "carouselUpFast",
    CarouselDown = "carouselDown",
    CarouselDownSlow = "carouselDownSlow",
    CarouselDownFast = "carouselDownFast",
    ShuffleTopRight = "shuffleTopRight",
    ShuffleTopRightSlow = "shuffleTopRightSlow",
    ShuffleTopRightFast = "shuffleTopRightFast",
    ShuffleRightTop = "shuffleRightTop",
    ShuffleRightTopSlow = "shuffleRightTopSlow",
    ShuffleRightTopFast = "shuffleRightTopFast",
    ShuffleRightBottom = "shuffleRightBottom",
    ShuffleRightBottomSlow = "shuffleRightBottomSlow",
    ShuffleRightBottomFast = "shuffleRightBottomFast",
    ShuffleBottomRight = "shuffleBottomRight",
    ShuffleBottomRightSlow = "shuffleBottomRightSlow",
    ShuffleBottomRightFast = "shuffleBottomRightFast",
    ShuffleBottomLeft = "shuffleBottomLeft",
    ShuffleBottomLeftSlow = "shuffleBottomLeftSlow",
    ShuffleBottomLeftFast = "shuffleBottomLeftFast",
    ShuffleLeftBottom = "shuffleLeftBottom",
    ShuffleLeftBottomSlow = "shuffleLeftBottomSlow",
    ShuffleLeftBottomFast = "shuffleLeftBottomFast",
    ShuffleLeftTop = "shuffleLeftTop",
    ShuffleLeftTopSlow = "shuffleLeftTopSlow",
    ShuffleLeftTopFast = "shuffleLeftTopFast",
    ShuffleTopLeft = "shuffleTopLeft",
    ShuffleTopLeftSlow = "shuffleTopLeftSlow",
    ShuffleTopLeftFast = "shuffleTopLeftFast",
    Zoom = "zoom"
}

export enum TransitionOutEnum {
    Fade = "fade",
    FadeSlow = "fadeSlow",
    FadeFast = "fadeFast",
    Reveal = "reveal",
    RevealSlow = "revealSlow",
    RevealFast = "revealFast",
    WipeLeft = "wipeLeft",
    WipeLeftSlow = "wipeLeftSlow",
    WipeLeftFast = "wipeLeftFast",
    WipeRight = "wipeRight",
    WipeRightSlow = "wipeRightSlow",
    WipeRightFast = "wipeRightFast",
    SlideLeft = "slideLeft",
    SlideLeftSlow = "slideLeftSlow",
    SlideLeftFast = "slideLeftFast",
    SlideRight = "slideRight",
    SlideRightSlow = "slideRightSlow",
    SlideRightFast = "slideRightFast",
    SlideUp = "slideUp",
    SlideUpSlow = "slideUpSlow",
    SlideUpFast = "slideUpFast",
    SlideDown = "slideDown",
    SlideDownSlow = "slideDownSlow",
    SlideDownFast = "slideDownFast",
    CarouselLeft = "carouselLeft",
    CarouselLeftSlow = "carouselLeftSlow",
    CarouselLeftFast = "carouselLeftFast",
    CarouselRight = "carouselRight",
    CarouselRightSlow = "carouselRightSlow",
    CarouselRightFast = "carouselRightFast",
    CarouselUp = "carouselUp",
    CarouselUpSlow = "carouselUpSlow",
    CarouselUpFast = "carouselUpFast",
    CarouselDown = "carouselDown",
    CarouselDownSlow = "carouselDownSlow",
    CarouselDownFast = "carouselDownFast",
    ShuffleTopRight = "shuffleTopRight",
    ShuffleTopRightSlow = "shuffleTopRightSlow",
    ShuffleTopRightFast = "shuffleTopRightFast",
    ShuffleRightTop = "shuffleRightTop",
    ShuffleRightTopSlow = "shuffleRightTopSlow",
    ShuffleRightTopFast = "shuffleRightTopFast",
    ShuffleRightBottom = "shuffleRightBottom",
    ShuffleRightBottomSlow = "shuffleRightBottomSlow",
    ShuffleRightBottomFast = "shuffleRightBottomFast",
    ShuffleBottomRight = "shuffleBottomRight",
    ShuffleBottomRightSlow = "shuffleBottomRightSlow",
    ShuffleBottomRightFast = "shuffleBottomRightFast",
    ShuffleBottomLeft = "shuffleBottomLeft",
    ShuffleBottomLeftSlow = "shuffleBottomLeftSlow",
    ShuffleBottomLeftFast = "shuffleBottomLeftFast",
    ShuffleLeftBottom = "shuffleLeftBottom",
    ShuffleLeftBottomSlow = "shuffleLeftBottomSlow",
    ShuffleLeftBottomFast = "shuffleLeftBottomFast",
    ShuffleLeftTop = "shuffleLeftTop",
    ShuffleLeftTopSlow = "shuffleLeftTopSlow",
    ShuffleLeftTopFast = "shuffleLeftTopFast",
    ShuffleTopLeft = "shuffleTopLeft",
    ShuffleTopLeftSlow = "shuffleTopLeftSlow",
    ShuffleTopLeftFast = "shuffleTopLeftFast",
    Zoom = "zoom"
}


// Transition
/** 
 * In and out transitions for a clip - i.e. fade in and fade out
**/
export class Transition extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=in" })
  in?: TransitionInEnum;

  @SpeakeasyMetadata({ data: "json, name=out" })
  out?: TransitionOutEnum;
}